insufficiently-tested attempt to stitch packets back together
When running with the Nautilus's Sonardyne, a single messages is
sometimes split across two socket.recv() calls. We were able to
reproduce this behavior on the bench using Moxas that weren't
configured properly. (turn off the 'a' delimiter)

This change was written during nui039, but we got out of range of
the Nautilus before I could test it.